editor får focus ved manuell endring av tekststil & clean up
