+++ skribenten-web/frontend/src/Brevredigering/LetterEditor/LetterEditor.tsx
@@ -6,32 +6,14 @@ import type { Dispatch, SetStateAction } from "react";
 import { createContext, useContext } from "react";
 
 import { DebugPanel } from "~/Brevredigering/LetterEditor/components/DebugPanel";
-import { applyAction, type CallbackReceiver } from "~/Brevredigering/LetterEditor/lib/actions";
+import { type CallbackReceiver } from "~/Brevredigering/LetterEditor/lib/actions";
 
-import Actions from "./actions";
 import { ContentGroup } from "./components/ContentGroup";
-import { EditorMenu, Typography } from "./components/EditorMenu";
+import { EditorMenu } from "./components/EditorMenu";
 import { SakspartView } from "./components/SakspartView";
 import { SignaturView } from "./components/SignaturView";
 import type { LetterEditorState } from "./model/state";
-
-const useEditorKeyboardShortcuts = (
-  editorState: LetterEditorState,
-  setEditorState: Dispatch<SetStateAction<LetterEditorState>>,
-) => {
-  return (event: React.KeyboardEvent<HTMLDivElement>) => {
-    if (event.altKey && event.code === "Digit1") {
-      event.preventDefault();
-      applyAction(Actions.switchTypography, setEditorState, editorState.focus, Typography.PARAGRAPH);
-    } else if (event.altKey && event.code === "Digit2") {
-      event.preventDefault();
-      applyAction(Actions.switchTypography, setEditorState, editorState.focus, Typography.TITLE1);
-    } else if (event.altKey && event.code === "Digit3") {
-      event.preventDefault();
-      applyAction(Actions.switchTypography, setEditorState, editorState.focus, Typography.TITLE2);
-    }
-  };
-};
+import { useEditorKeyboardShortcuts } from "./utils";
 
 export const LetterEditor = ({
   freeze,
