+++ skribenten-web/frontend/src/Brevredigering/LetterEditor/editor.css
@@ -1,7 +1,7 @@
 .editor {
     margin-top: var(--a-spacing-6);
     /*clamp: minimum, ideal, max */
-    padding: 0 clamp(2rem, 6vw, 8rem);
+    padding: 0 clamp(1rem, 2vw, 4rem);
     caret-color: green;
 
     [contenteditable] {
